+++ uikit-samples/src/main/java/com/sendbird/uikit/samples/common/extensions/ViewExtentions.kt
@@ -3,8 +3,12 @@ package com.sendbird.uikit.samples.common.extensions
 import android.content.Context
 import android.os.Build
 import android.view.View
+import android.view.Window
 import android.widget.TextView
 import androidx.annotation.ColorRes
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowCompat
+import androidx.core.view.WindowInsetsCompat
 
 @Suppress("DEPRECATION")
 internal fun TextView.setAppearance(context: Context, res: Int) {
@@ -22,3 +26,20 @@ internal fun TextView.setTextColorResource(@ColorRes id: Int) {
 internal fun View.toggleVisibility() {
     visibility = if (visibility == View.VISIBLE) View.GONE else View.VISIBLE
 }
+
+internal fun View.setInsetMargin(window: Window) {
+    // For custom edge-to-edge system bar color
+//    val edgeToEdgeConfig = EdgeToEdgeConfig(
+//        statusBarColorLight = context.getColorResource(R.color.primary_300),
+//        statusBarColorDark = context.getColorResource(R.color.primary_100)
+//    )
+//    SendbirdUIKit.setEdgeToEdgeConfig(edgeToEdgeConfig)
+
+    ViewCompat.setOnApplyWindowInsetsListener(this) { v, windowInsets ->
+        val insets = windowInsets.getInsets(WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.ime())
+        v.setPadding(insets.left, insets.top, insets.right, insets.bottom)
+        WindowCompat.getInsetsController(window, window.decorView).isAppearanceLightStatusBars = true
+        WindowInsetsCompat.CONSUMED
+    }
+}
+