+++ uikit/src/main/java/com/sendbird/uikit/internal/ui/messages/OtherMessageView.kt
@@ -11,6 +11,7 @@ import com.sendbird.android.message.SendingStatus
 import com.sendbird.uikit.R
 import com.sendbird.uikit.consts.MessageGroupType
 import com.sendbird.uikit.databinding.SbViewOtherMessageComponentBinding
+import com.sendbird.uikit.internal.extensions.isNewLineMessage
 import com.sendbird.uikit.internal.extensions.toContextThemeWrapper
 import com.sendbird.uikit.model.MessageListUIParams
 import com.sendbird.uikit.model.configurations.ChannelConfig
@@ -76,6 +77,7 @@ internal class OtherMessageView @JvmOverloads internal constructor(
 
         binding.ivProfileView.visibility = if (showProfile) VISIBLE else INVISIBLE
         binding.tvNickname.visibility = if (showNickname) VISIBLE else GONE
+        binding.newLineView.visibility = if (message.isNewLineMessage) VISIBLE else GONE
         binding.emojiReactionListBackground.visibility =
             if (enableReactions) VISIBLE else GONE
         binding.rvEmojiReactionList.visibility = if (enableReactions) VISIBLE else GONE